requirements.txt
Django==4.2.6
djangorestframework==3.14.0
drf-action-serializer==1.0.5
drf-flex-fields==1.0.2
